@@ -1,6 +1,6 @@
 import BezierEasing from "bezier-easing";
-import anime from "animejs";
-import { workScrollSpeed } from "./store";
+import { animate, stagger } from "animejs";
+import { workScrollState } from "$lib/state.svelte";
 import { quintOut } from "svelte/easing";
 
 // Intro Letter reveal animation used declaratively with the 'in:' svelte directive or programmatically with anime.js
@@ -74,14 +74,13 @@ export function letterSlideIn (node: HTMLElement, params?: { duration?: number, 
     // Call animation programmatically outside of svelte transitions with anime.js
     function animeAnimation(animeParams?: { easing?: string, onComplete?: () => void }) {
         if (animeParams === undefined) animeParams = {};
-        anime({
-            targets: animeTargets,
+        animate(animeTargets, {
             translateX: "0%",
-            easing: (animeParams.easing) ? animeParams.easing : "cubicBezier(.2, .58, .43, 1)",
-            duration: params!.duration,
-            delay: anime.stagger(params!.delay!, { start: params!.initDelay }),
+            duration: params!.duration!,
+            ease: (animeParams.easing) ? animeParams.easing : "cubicBezier(.2, .58, .43, 1)",
+            delay: stagger(params!.delay!, { start: params!.initDelay }),
             // Return node to original state if destroyLettersUponSuccess is true
-            complete: () => {
+            onComplete: () => {
                 // Return node to original state on completion
                 node.innerHTML = originalNodeHTML;
                 if (animeParams)
@@ -110,6 +109,7 @@ export function letterSlideOut (node: HTMLElement, params?: { duration?: number,
 
     // Wrap every letter with a div and prepare it for animation
     let masks = tagLettersAndWords(node, { breakWord: params.breakWord });
+
     // Set initial style properties for each mask before animation begins
     masks.forEach((e: HTMLElement) => {
         e.style.display = "inline-flex";
@@ -128,29 +128,32 @@ export function letterSlideOut (node: HTMLElement, params?: { duration?: number,
         duration: params.duration,
         // Svelte transition animation
         tick: (t: number) => {
-            let eased = BezierEasing(.32, .24, .76, .26)(t);
+            const eased = BezierEasing(.32, .24, .76, .26)(t);
 
             masks.forEach((element: HTMLElement) => {
-                element.childNodes.forEach(child => {
-                    (child as HTMLElement).style.transform = `translate3d(${(-150 + (eased * 150)).toFixed(2)}%, 0px, 0px)`;
+                // Letter animation
+                element.childNodes.forEach(e => {
+                    (e as HTMLElement).style.transform = `translate3d(${(-150 + (eased * 150)).toFixed(2)}%, 0px, 0px)`;
                 });
 
+                // Word animation
                 element.style.transform = `translate3d(${(-80 + (eased * 80)).toFixed(2)}%, 0px, 0px)`;
             });
 
-            if (eased >= 1) node.innerHTML = originalNodeHTML;
+            if (eased <= 0) {
+                node.innerHTML = originalNodeHTML;
+            }
         },
 
         // Call animation outside of svelte blocks programmatically with anime.js
         anime: (easing?: string) => {
-            anime({
-                targets: animeTargets,
+            animate(animeTargets, {
                 translateX: "-150%",
-                easing: easing ? easing : "cubicBezier(.2, .58, .43, 1)",
-                duration: params!.duration,
-                delay: anime.stagger(params!.delay!, { start: params!.initDelay }),
+                ease: easing ? easing : "cubicBezier(.2, .58, .43, 1)",
+                duration: params!.duration!,
+                delay: stagger(params!.delay!, { start: params!.initDelay }),
                 // Return node to original state if destroyLettersUponSuccess is true
-                complete: () => {
+                onComplete: () => {
                     // Return node to original state on completion
                     node.innerHTML = originalNodeHTML;
                 }
@@ -236,12 +239,11 @@ export function maskSlideIn(node: HTMLElement, params?: { duration?: number, del
 
     // Call animation programmatically outside of svelte transitions with anime.js
     function animeAnimation(easing?: string) {
-        anime({
-            targets: [mask, node],
+        animate([mask, node], {
             translateX: "0%",
-            easing: easing ? easing : "cubicBezier(.58,.14,.06,.97)",
-            duration: params!.duration,
-            delay: params!.delay
+            ease: easing ? easing : "cubicBezier(.58,.14,.06,.97)",
+            duration: params!.duration!,
+            delay: params!.delay!
         })
     }
 }
@@ -285,13 +287,12 @@ export function workImageIntro(node: HTMLElement, params?: { promise: Promise<an
     node.style.marginRight = "60%";
 
     params.promise.then(() => {
-        anime({
-            targets: node,
+        animate(node, {
             marginRight: "0%",
-            easing: "easeOutQuint",
+            ease: "outQuint",
             duration: 1400,
-            delay: params.delay,
-            complete: () => {
+            delay: params.delay!,
+            onComplete: () => {
                 node.style.marginRight = "";
                 node.style.transition = "";
             }
@@ -313,17 +314,16 @@ export function workListIntro(node: HTMLElement, params?: { promise: Promise<any
     node.style.transform = "translateX(100%)"
 
     params.promise.then(() => {
-        anime({
-            targets: node, 
+        animate(node, {
             translateX: "0%",
-            easing: "easeOutQuint",
+            ease: "outQuint",
             duration: 1600,
-            delay: params.delay,
-            update: (anim) => {
-                const t = 1 - quintOut(anim.progress / 100);
-                workScrollSpeed.set(t * 2500);
+            delay: params.delay!,
+            onUpdate: (anim: { progress: number; }) => {
+                const t = 1 - quintOut(anim.progress);
+                workScrollState.speed = t * 2500;
             },
-            complete: () => {
+            onComplete: () => {
                 node.style.transform = "";
                 node.style.transition = "";
                 if (params.onComplete) params.onComplete();
