@@ -2,19 +2,20 @@
 
     import { onMount } from "svelte";
     import { letterSlideIn, maskSlideIn } from "$lib/animations";
-    import { loadPagePromise, siteDataFetch } from "$lib/store";
+    import { loadPagePromise } from "$lib/store";
     import { onScrolledIntoView } from "$lib/utils";
-    import type { SiteData } from "$lib/types";
+    import { dataState } from "$lib/state.svelte";
 
-    let footerContainerElement: HTMLElement, logoElement: HTMLElement, creditsElement: HTMLElement, statusElement: HTMLElement, fullEmailLinkElement: HTMLElement;
-    let signaturePath1: SVGPathElement, signaturePath2: SVGPathElement, signaturePath3: SVGPathElement, signaturePath4: SVGPathElement; 
+    let footerContainerElement: HTMLElement = $state()!
+    let logoElement: HTMLElement = $state()!; 
+    let creditsElement: HTMLElement = $state()!; 
+    let statusElement: HTMLElement = $state()!; 
+    let fullEmailLinkElement: HTMLElement = $state()!;
 
-    let siteData: SiteData = { availablity_date: "" };
-
-    siteDataFetch.subscribe(val => {
-        if (val !== undefined)
-            siteData = val;
-    });
+    let signaturePath1: SVGPathElement = $state()!; 
+    let signaturePath2: SVGPathElement = $state()!; 
+    let signaturePath3: SVGPathElement = $state()!;
+    let signaturePath4: SVGPathElement = $state()!; 
 
     const currentYear = new Date().getFullYear();
     
@@ -83,16 +84,18 @@
         </div>
 
         <div class="status-wrapper">
-                {#if siteData.availablity_date === ""}
+            {#if dataState.siteData}
+                {#if dataState.siteData!.availablity_date === ""}
                     <p class="large-text" bind:this={statusElement}>
                         i am currently accepting freelance work, <br>you may reach me on my email.
                     </p>
                 {:else}
                     <p class="large-text" bind:this={statusElement}>
-                        i am available for freelance work after <br> {siteData.availablity_date}.
+                        i am available for freelance work after <br> {dataState.siteData.availablity_date}.
                     </p>
                 {/if}
-            <a class="button large-text" bind:this={fullEmailLinkElement} href="mailto:musabhassan04@gmail.com" target="_blank">musabhassan04@gmail.com</a>
+            {/if}
+            <a class="button large-text" bind:this={fullEmailLinkElement} href="mailto:musab@musabhassan.com" target="_blank">musab@musabhassan.com</a>
         </div>
         
         <div class="credits-wrapper" bind:this={creditsElement}>
@@ -142,8 +145,9 @@
 
 <style lang="sass">
 
-@import "../consts.sass"
-@include textStyles()
+@use "../consts.sass" as consts
+
+@include consts.textStyles()
 
 .footer-wrapper
     width: 100vw
@@ -189,7 +193,7 @@
 
         p.year
             margin-bottom: 1vh
-            font-family: $font
+            font-family: consts.$font
             font-size: 1.8vh
             font-weight: normal
             color: rgba(255,255,255,0.3)
